Moved test_helper outside the tests folder
